executive_dashboard/requirements.txt
# Executive Dashboard V4.0 Dependencies
# Install with: pip install -r requirements.txt

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# JWT and authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

# HTTP client for API calls
httpx>=0.26.0

# Logging
python-logging>=0.1.0

# Environment variables
python-dotenv>=1.0.0
